+++ index-cosmetics.ts
@@ -1,2 +1,1 @@
 export { default as CosmeticsInjection } from './src/cosmetics-injection';
-export { overrideUserAgent } from './src/cosmetics-injection';