+++ index.ts
@@ -1,6 +1,5 @@
 // Cosmetic injection
 export { default as CosmeticsInjection } from './src/cosmetics-injection';
-export { overrideUserAgent } from './src/cosmetics-injection';
 
 // Blocking
 export { default as FiltersEngine } from './src/engine/engine';
@@ -8,8 +7,8 @@ export { default as ReverseIndex } from './src/engine/reverse-index';
 export { default as Request } from './src/request';
 export { deserializeEngine } from './src/serialization';
 
-export {default as matchCosmeticFilter } from './src/matching/cosmetics';
-export {default as matchNetworkFilter } from './src/matching/network';
+export { default as matchCosmeticFilter } from './src/matching/cosmetics';
+export { default as matchNetworkFilter } from './src/matching/network';
 
 export { parseCosmeticFilter } from './src/parsing/cosmetic-filter';
 export { parseNetworkFilter } from './src/parsing/network-filter';